LayoutTests/fast/forms/time-multiple-fields/time-multiple-fields-wheel-event-expected.txt

Test for wheel operations for <input type=time>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial value is 07:00. We'll wheel up by 08:00:
PASS input.value is "08:00"
Wheel up by 100:
PASS input.value is "09:00"
Wheel down by 1:
PASS input.value is "08:00"
Wheel down by 256:
PASS input.value is "07:00"
Disabled input element:
PASS input.value is "07:00"
Read-only input element:
PASS input.value is "07:00"
No focus:
PASS input.value is "07:00"
PASS successfullyParsed is true

TEST COMPLETE
